added copyIdButton to showAnnotation
While debugging a set of bad annos today I needed to collect a subset
of their identifiers to enter into a script. This button serves a dual
role of making the annotation id visible and making making it possible
to copy the id to the clipboard with a single click. The value can be
passed in as a parameter so the share link, direct link, etc could be
copied rather than the raw id, depending on what a workflow calls for.